6.1.12 Exclude /var/lib/docker/overlay2
Docker overlay contains files which are owned by groups inside
the various Docker containers present on the system. Although
CIS 6.1.12 rightfully considers this a security concern, requiring
the main system has a complete mapping of all groups inside the
docker containers located on the system is undesirable.

Similar issue exists for LXD Containers, and with the same motivation
these data directories are excluded from the analysis.

Signed-off-by: Ivo van Doorn <ivovandoorn@samotics.com>